Updated getAssignmentStatements and README.txt
Doodled a bit in the comments for getAssignmentStatements.

Also, I'm am going to use the README.txt to record down our thoughts and
suggestions on the various parts, and what we have implemented. That
way, we can just check the text file rather than calling or texting each
other when we are at home.