Updated advancedParserClasses
Created two private boolean variables that will be set to true if a
syntax or a logic error is found.

checkSecondHalf will receive a string and a char. The char will be the
first half of a pair of punctuators. The function will return true if it
has found the second half. Otherwise, it returns false.