Fixed advancedParserClasses.pp and project.cpp
Tidied up the code in the file, and fixed a couple of bugs.

When we call values from other classes, we need to use its namespace.
For example, for numElement_tableOfKeywords, we need to code
'ensc251::numElement_tableOfKeywords'.

In project.cpp, we check to see if the string is empty or not. If it is,
we just skip it (blank space).