Update advancedParserClasses.cpp
Got rid of class Identify. Clippy will handle the entire process of part
1-3 (I placed Identify functions into private).

We will call clippy just after removing all the comments. In it, it will
perform get the assignment statements and function declarations.

When I tried to run it, I found a bug where I could not get pass
parserclass. It has something to do with the string position.