Major Changes to everything
In parserClasses, we now set the enum to stringClass rather than
stringType. stringType variables are solely reserved for class types
(stringClass is to stringType like how identifeir is to integers). I
fixed these recurrances in all of parserClasses and project.cpp.

In advancedParserClasses, I finished implementing the
setIdentifierClass. Basically, we look at each token. If a token is an
identifier, we take the string of the current token and set it to the
next tokne's stringType.

I am going to look into the bug that prevents us from going through
parserClasses now.