Started on checkSecondHalf
checkSecondHalf is a recursive function that will receive (maybe) a
tokenList and the previous number of errors.

If it finds a beginning bracket (e.g '(' ), the number of errors is
incremented. If it finds an end bracket (e.g ')' ), the number of errors
is decremented.

Ideally, then it will go into a recursive function to keep carrying on.
Stopping case is when the getNext produces a NULL.