Update project.cpp
Modified getAssignment statements functions. I read posts about it on piazza and assignment statements as explicitly defined as
"variable = expression;" and our function just kept appending the token until the end of the statement which is incorrect. For example 
(int i = 0) is not an assignment statement as there is no ";" after the zero.