Update project.cpp
I added functions that calculates and stores the statistics of the tokens in the input file. Also included a prompt for the user to choose whether they want to use the program in non-verbose,verbose, traverse through the list and get tokens(mostly used for testing may remove in final version), or exit the program. I ended up throwing everything in the project file as it was easier and I couldnt figure out a better way to do so in our limited time. So i didn't end up using the advancedParser files for this but the code there was helpful. There is a bug with my getFunctionDeclarations so i will work on that.