Added some comments and improved checkSecondHalf
I added some comments relating to the entire project. Feel free to check
README.txt.

For checkSecondHalf, we compare the number of left and right
parenthesis. When the next token is NULL, it will output if any errors
are found (using compareTwo function). After that, we set the left to
equal to the right.